Implement alpha-beta (#3)
Implement alpha-beta negamax. This also removes hope-chess, and commits
to the One True Engine.

Changes:

- PERF: enable optimizations in dev profile as well (otherwise too slow)
- PERF: Add config for Cargo to use native CPU instructions when
possible
- PERF: change from f32 to i64 for evaluation (I don't know if this
actually improves anything)
- REFACTOR: update material evaluation to be split up more and easier to
read
- FEATURE: include mobility as a factor in evaluation
- FEATURE: implement alpha-beta negamax
- REFACTOR: remove hope chess algorithm

It is sufficiently strong now to beat me and my wife. We're ~1400 on
chess.com, so it's a decent strength but can clearly get better, so
future improvements will be fun. I think this will require implementing
UCI so that I can pit it against engines.

+++ src/evaluation/types.rs
@@ -1,1 +1,1 @@
-pub type Score = f32;
+pub type Score = i64;
